+++ components/Timeline/TimelinePage.tsx
@@ -3,7 +3,7 @@
 import { useTimelineProvider } from "@/providers/TimelineProvider";
 import TimelineHero from "@/components/Timeline/TimelineHero";
 import MobileMomentsSection from "@/components/Timeline/MobileMomentsSection";
-import { useInProcessMomentsSocket } from "@/hooks/useInProcessMomentsSocket";
+import { useInProcessMomentsChannel } from "@/hooks/useInProcessMomentsChannel";
 import useIsMobile from "@/hooks/useIsMobile";
 import MobileTimeline from "./MobileTimeline";
 import TimelineMobileMoon from "./TimelineMobileMoon";
@@ -12,7 +12,7 @@ import TimelineSpiral from "./TimelineSpiral";
 
 const TimelinePage = () => {
   const { error } = useTimelineProvider();
-  useInProcessMomentsSocket();
+  useInProcessMomentsChannel();
   const isMobile = useIsMobile();
 
   if (error) return <main>Error loading timeline.</main>;
