+++ providers/MomentProvider.tsx
@@ -3,7 +3,7 @@
 import { Moment } from "@/types/moment";
 import { createContext, useContext, ReactNode } from "react";
 import useMomentData from "@/hooks/useMomentData";
-import useMomentSocket from "@/hooks/useMomentSocket";
+import useMomentChannel from "@/hooks/useMomentChannel";
 
 const MomentContext = createContext<
   | (ReturnType<typeof useMomentData> & {
@@ -14,7 +14,7 @@ const MomentContext = createContext<
 
 export function MomentProvider({ children, moment }: { children: ReactNode; moment: Moment }) {
   const momentdata = useMomentData(moment);
-  useMomentSocket(moment, momentdata.fetchMomentData);
+  useMomentChannel(moment, momentdata.fetchMomentData);
 
   return (
     <MomentContext.Provider
